#1012 create etl_table.py to hold table constructors
