+++ src/ch18_world_etl/etl_sqlstrs.py
@@ -6,111 +6,7 @@ from src.ch01_py.db_toolbox import (
 from src.ch17_idea.idea_config import get_idea_config_dict, get_quick_ideas_column_ref
 from src.ch17_idea.idea_db_tool import create_idea_sorted_table, get_default_sorted_list
 from src.ch18_world_etl._ref.ch18_semantic_types import KnotTerm
-
-ALL_DIMEN_ABBV7 = {
-    "MMTPAYY",
-    "MMTBUDD",
-    "MMTHOUR",
-    "MMTMONT",
-    "MMTWEEK",
-    "MMTOFFI",
-    "MMTUNIT",
-    "BLFMEMB",
-    "BLFVOCE",
-    "BLFAWAR",
-    "BLFFACT",
-    "BLFHEAL",
-    "BLFCASE",
-    "BLFREAS",
-    "BLFLABO",
-    "BLFPLAN",
-    "BLFUNIT",
-    "NBUEPCH",
-    "TRLTITL",
-    "TRLNAME",
-    "TRLROPE",
-    "TRLLABE",
-}
-
-
-def get_dimen_abbv7(dimen: str) -> str:
-    return {
-        "moment_paybook": "MMTPAYY",
-        "moment_budunit": "MMTBUDD",
-        "moment_epoch_hour": "MMTHOUR",
-        "moment_epoch_month": "MMTMONT",
-        "moment_epoch_weekday": "MMTWEEK",
-        "moment_timeoffi": "MMTOFFI",
-        "momentunit": "MMTUNIT",
-        "belief_voice_membership": "BLFMEMB",
-        "belief_voiceunit": "BLFVOCE",
-        "belief_plan_awardunit": "BLFAWAR",
-        "belief_plan_factunit": "BLFFACT",
-        "belief_plan_healerunit": "BLFHEAL",
-        "belief_plan_reason_caseunit": "BLFCASE",
-        "belief_plan_reasonunit": "BLFREAS",
-        "belief_plan_partyunit": "BLFLABO",
-        "belief_planunit": "BLFPLAN",
-        "beliefunit": "BLFUNIT",
-        "nabu_epochtime": "NBUEPCH",
-        "translate_title": "TRLTITL",
-        "translate_name": "TRLNAME",
-        "translate_rope": "TRLROPE",
-        "translate_label": "TRLLABE",
-        "translate_core": "TRLCORE",
-    }.get(dimen)
-
-
-class prime_tablenameException(Exception):
-    pass
-
-
-def create_prime_tablename(
-    idea_dimen_or_abbv7: str, phase: str, stage: str, put_del: str = None
-) -> str:
-    """
-    phase must be one: 's', 'h', 'job'
-    stage must be one: 'raw', 'agg', 'vld'
-    """
-
-    abbv_references = {
-        "MMTPAYY": "moment_paybook",
-        "MMTBUDD": "moment_budunit",
-        "MMTHOUR": "moment_epoch_hour",
-        "MMTMONT": "moment_epoch_month",
-        "MMTWEEK": "moment_epoch_weekday",
-        "MMTOFFI": "moment_timeoffi",
-        "MMTUNIT": "momentunit",
-        "BLFMEMB": "belief_voice_membership",
-        "BLFVOCE": "belief_voiceunit",
-        "BLFAWAR": "belief_plan_awardunit",
-        "BLFFACT": "belief_plan_factunit",
-        "BLFGROU": "belief_groupunit",
-        "BLFHEAL": "belief_plan_healerunit",
-        "BLFCASE": "belief_plan_reason_caseunit",
-        "BLFREAS": "belief_plan_reasonunit",
-        "BLFLABO": "belief_plan_partyunit",
-        "BLFPLAN": "belief_planunit",
-        "BLFUNIT": "beliefunit",
-        "NBUEPCH": "nabu_epochtime",
-        "TRLTITL": "translate_title",
-        "TRLNAME": "translate_name",
-        "TRLROPE": "translate_rope",
-        "TRLLABE": "translate_label",
-        "TRLCORE": "translate_core",
-    }
-    tablename = idea_dimen_or_abbv7
-    if abbv_references.get(idea_dimen_or_abbv7.upper()):
-        tablename = abbv_references.get(idea_dimen_or_abbv7.upper())
-    if phase in {"s", "h", "job"}:
-        tablename = f"{tablename}_{phase}"
-    if stage is None:
-        return tablename
-    if stage not in {"raw", "agg", "vld"}:
-        raise prime_tablenameException(f"'{stage}' is not a valid stage")
-
-    return f"{tablename}_{put_del}_{stage}" if put_del else f"{tablename}_{stage}"
-
+from src.ch18_world_etl.etl_table import create_prime_tablename
 
 CREATE_TRLTITL_SOUND_RAW_SQLSTR = """CREATE TABLE IF NOT EXISTS translate_title_s_raw (idea_number TEXT, spark_num INTEGER, face_name TEXT, otx_title TEXT, inx_title TEXT, otx_knot TEXT, inx_knot TEXT, unknown_str TEXT, error_message TEXT)"""
 CREATE_TRLTITL_SOUND_AGG_SQLSTR = """CREATE TABLE IF NOT EXISTS translate_title_s_agg (spark_num INTEGER, face_name TEXT, otx_title TEXT, inx_title TEXT, otx_knot TEXT, inx_knot TEXT, unknown_str TEXT, error_message TEXT)"""
